feat:add basic components
