+++ src/components/basic/TweetUserName2.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+const TweetUserName2 = ({name, account})=>{
+    return (
+        <div className='tweet-username-2'>
+          <span className='user-name'>
+            {name}
+          </span>
+          <span className='user-account'>
+            @{account}
+          </span>
+        </div>
+    )
+  }
+
+  export default TweetUserName2;
